pyproject.toml
[project]
name = "mos"
version = "0.1.0"
description = "Moradabad AI Design + Production OS"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6,<3",
    "sqlalchemy>=2.0,<3",
    "alembic>=1.13,<2",
    "psycopg[binary]>=3.1,<4",
    "pydantic-settings>=2.2,<3",
    "cadquery>=2.4,<3",
    "trimesh>=4.0,<5",
    "numpy>=1.26,<3",
    "pyyaml>=6.0,<7",
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.27,<1",
    "celery>=5.3,<6",
    "boto3>=1.34,<2",
    "httpx>=0.27,<1",
    "reportlab>=4.0,<5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-cov>=4.1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra --strict-markers"
